refactor: dto - request, response 하위 폴더 분류
