+++ backend/src/main/java/backend/mulkkam/auth/dto/response/OauthLoginResponse.java
@@ -1,4 +1,4 @@
-package backend.mulkkam.auth.dto;
+package backend.mulkkam.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
